Material widget

+++ SofaImGui/src/SofaImGui/ImGuiGUIEngine.cpp
@@ -772,7 +772,7 @@ void ImGuiGUIEngine::beforeDraw(GLFWwindow*)
             m_currentFBOSize = {static_cast<unsigned int>(m_viewportWindowSize.first), static_cast<unsigned int>(m_viewportWindowSize.second)};
         }
     }
-    sofa::core::visual::VisualParams::defaultInstance()->viewport() = {0,0,m_currentFBOSize.first, m_currentFBOSize.second};
+    sofa::core::visual::VisualParams::defaultInstance()->viewport() = {0,0, static_cast<int>(m_currentFBOSize.first), static_cast<int>(m_currentFBOSize.second)};
 
     m_fbo->start();
 }
